feat: Move codes to check permissions at the right time
