+++ app/src/main/AndroidManifest.xml
@@ -32,15 +32,6 @@
                 android:resource="@xml/provider_paths" />
         </provider>
 
-<!--        <activity android:name=".SelectUser">-->
-<!--            <intent-filter>-->
-<!--                <action android:name="android.intent.action.MAIN" />-->
-
-<!--                <category android:name="android.intent.category.LAUNCHER" />-->
-<!--            </intent-filter>-->
-<!--        </activity>-->
-
-
         <activity
             android:name=".Splash"
             android:screenOrientation="portrait">
