+++ app/src/main/java/com/example/yaneodoo/Splash.java
@@ -1,16 +1,11 @@
 package com.example.yaneodoo;
 
-import android.Manifest;
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.os.Handler;
-import android.text.TextUtils;
 import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
 
 public class Splash extends AppCompatActivity {
 
@@ -22,7 +17,6 @@ public class Splash extends AppCompatActivity {
 
         Handler hd = new Handler();
         hd.postDelayed(new splashhandler(), 2000); // 1초 후에 hd handler 실행  3000ms = 3초
-
     }
 
     private class splashhandler implements Runnable{
@@ -30,35 +24,6 @@ public class Splash extends AppCompatActivity {
             startActivity(new Intent(getApplication(), Login.class)); //로딩이 끝난 후, ChoiceFunction 이동
             Splash.this.finish(); // 로딩페이지 Activity stack에서 제거
         }
-    }
-
-
-    public void checkSelfPermission() {
-        String temp = "";
-
-        //카메라 권한 확인
-        if (ContextCompat.checkSelfPermission(this,
-                Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
-            temp += Manifest.permission.CAMERA + " ";
-        }
-
-        //파일 쓰기 권한 확인
-        if (ContextCompat.checkSelfPermission(this,
-                Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
-            temp += Manifest.permission.WRITE_EXTERNAL_STORAGE + " ";
-        }
-
-        //파일 읽기 권한 확인
-        if (ContextCompat.checkSelfPermission(this,
-                Manifest.permission.READ_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
-            temp += Manifest.permission.READ_EXTERNAL_STORAGE + " ";
-        }
-
-        //권한 요청
-        if (TextUtils.isEmpty(temp) == false)
-            ActivityCompat.requestPermissions(this, temp.trim().split(" "), 1);
-        else
-            Toast.makeText(this, "권한을 모두 허용", Toast.LENGTH_SHORT).show();
     }
 
     // 마지막으로 뒤로 가기 버튼을 눌렀던 시간 저장
